// src/enum/auditoriaUsuario.enum.ts
export enum AUDITORIAUSUARIO_ENUM {
  CREACION = "CREACION",
  ACTUALIZACION = "ACTUALIZACION",
  ELIMINACION = "ELIMINACION",
  DESACTIVACION = "DESACTIVACION",
  ACTIVACION = "ACTIVACION",
  ACTUALIZACION_DE_PERMISOS = "ACTUALIZACION_DE_PERMISOS",
  APROBAR_SOLICITUD = "APROBAR_SOLICITUD",
  DENEGAR_SOLICITUD = "DENEGAR_SOLICITUD",
  TRANSFERENCIA = "TRANSFERENCIA",
}